requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cloudinary==1.24.0
cryptography==3.3.1
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==14.2.0
django-braces==1.14.0
django-crispy-forms==1.9.0
django-decouple==2.1
django-heroku==0.3.1
django-model-utils==4.1.1
django-star-ratings==0.9.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.9
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.5
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-core==4.0.3
sqlparse==0.4.1
swapper==1.1.2.post1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
userprofile==0.2.5
whitenoise==5.0.1
